handle updating field manual

+++ CORE/MANUAL/50_MANUAL.sh
@@ -17,7 +17,14 @@ LOGFILE=$MODULE_DIR/$MODULE.log
 ### MODULE COMMANDS FUNCTION ###
 module_commands () {
 
-sudo cp $MODULE_DIR/arcOS-Field-Manual.html /opt/arcOS/
+SAVE_DIR=$ARCOS_DATA/QRV/$MYCALL/SAVED/$MODULE
+mkdir -p $SAVE_DIR
+
+if [ ! -f $SAVE_DIR/arcOS-Field-Manual.html ]; then
+    sudo cp $MODULE_DIR/arcOS-Field-Manual.html /opt/arcOS/
+else
+    sudo cp $SAVE_DIR/arcOS-Field-Manual.html /opt/arcOS/
+fi
 cp $MODULE_DIR/arcOS-Field-Manual.desktop $HOME/.local/share/applications/
 
 } # END OF MODULE COMMANDS FUNCTION
